+++ bootique-kafka-client/src/main/java/io/bootique/kafka/client/consumer/ConsumerFactory.java
@@ -135,6 +135,4 @@ public class ConsumerFactory {
 
         return new KafkaConsumer<>(properties, config.getKeyDeserializer(), config.getValueDeserializer());
     }
-
-
 }